# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Atualizado em 21/03/2021 02:47",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticação",
                "description": "Autenticação",
                "operationId": "authLogin",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/auth/me": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Dados do usuário logado",
                "description": "Dados do usuário logado",
                "operationId": "authMe",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sair do sistema",
                "description": "Sair do sistema",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Atualiza token",
                "description": "Atualiza token",
                "operationId": "authRefresh",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/auth/password-token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inicia o processo de alteração de senha enviando e-mail com token para usuário",
                "description": "Inicia o processo de alteração de senha enviando e-mail com token para usuário",
                "operationId": "authPasswordToken",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/auth/password-refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Finaliza processo de alteração de senha recebendo e-mail, token, nova senha e confirmação de nova senha",
                "description": "Finaliza processo de alteração de senha recebendo e-mail, token, nova senha e confirmação de nova senha",
                "operationId": "authPasswordRefresh",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/api/setting/search": {
            "get": {
                "tags": [
                    "Setting"
                ],
                "summary": "Buscar Settings",
                "description": "Buscar Settings",
                "operationId": "setting-search",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Quantidade de resultados por página",
                        "required": false
                    },
                    {
                        "name": "orderby",
                        "in": "query",
                        "description": "Nome da coluna de referência para ordenação",
                        "required": false
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordem asc ou desc",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/setting/find/{id}": {
            "get": {
                "tags": [
                    "Setting"
                ],
                "summary": "Retornar Setting por id",
                "description": "Retornar Setting por id",
                "operationId": "setting-find",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de order",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados de order"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/setting/save": {
            "post": {
                "tags": [
                    "Setting"
                ],
                "summary": "Salvar Settings",
                "description": "Salvar Settings",
                "operationId": "setting-save",
                "responses": {
                    "200": {
                        "description": "Dados salvos"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/setting/delete/{id}": {
            "post": {
                "tags": [
                    "Setting"
                ],
                "summary": "Deletar Setting",
                "description": "Deletar Setting",
                "operationId": "setting-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de order",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados salvos"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}